regalloc: fix overwriting of registers
When a register was reallocated, it can overwrite the register values of
other registers. For example, when register A is alive before and after
register B. Then when register A is reallocated after deallocating
register B, then register A gets the same register as B. However, A and

B should have different register values. This commit fixes this issue by
simply checking which machine registers overlap with the current
register and using the first machine register that does not overlap with
the current interval. This completely removes reallocation and
simplifies the register allocator.