use linked list of CALL instructions for args
A call instruction which requires n arguments now consists of n + 1 call
instructions, where the first specifices the called function and points
to the next call instruction. Each successive call instruction is a
parameter which points to the next instruction. The instructions are
specified in reverse order for easier code generation.

In the future, there could be an IR pass which reverses the order of the
function call parameters.