fix declaration post-processing
The code adds a declaration to the scope. However, this should only
happen if the declaration does not consist of a type. Previously, this
meant that the declaration does not have a name and contains an
identifier. However, now a declaration without a name is just a type
node, not a declaration. This fixes the code for that.